+++ src/el/ctk_utils/__init__.py
@@ -18,6 +18,5 @@ from ._colors import (
     tk_to_rgb8,
     rgb_to_hex_str,
 )
-from ._tooltip import CTkToolTip
 from ._dynamic_config import apply_to_config, apply_to_tk_var
 from ._scaling import apply_widget_scaling